# src/main/resources/v1/eric-oss-auto-config-consistency-openapi.yaml
...
openapi: "3.0.2"
servers:
  - url: http://eacc/v1
info:
  x-api-id: 8608e72f-121d-4049-be33-7eb2d3cf5594
  x-audience: external-public
  description: |
    Ericsson Automated Configuration Consistency (EACC) provides self-service analytics 
    to enable users to uncover baseline consistency errors. EACC supports the scheduling 
    of CM consistency audits which detect inconsistencies based on user-supplied rules. 
    These inconsistencies can be corrected manually by the user. 
    Any changes can also be manually rolled back by the user.
  version: 0.1.5
  title: Ericsson Automated Configuration Consistency
  license:
    name: COPYRIGHT Ericsson 2024
    url: https://www.ericsson.com/en/legal
  termsOfService: https://www.ericsson.com/en/legal
  contact:
    email: pdlonteamq@pdl.internal.ericsson.com
...
paths:
  /v1/rulesets:
    get:
      tags:
        - rulesets
      summary: Get metadata for all Rulesets.
      description: Get metadata for all Rulesets. An empty list is returned if no Rulesets exist.
      operationId: listRulesets
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RulesetMetadata'
              example:
                [
                  {
                    "id": "e0399094-0b34-4ccd-bcf7-a904293594c5",
                    "rulesetName": "ruleset1",
                    "uri": "v1/rulesets/e0399094-0b34-4ccd-bcf7-a904293594c5"
                  },
                  {
                    "id": "fcbd87fd-8586-4588-926c-d2d6f9dfd584",
                    "rulesetName": "ruleset2",
                    "uri": "v1/rulesets/fcbd87fd-8586-4588-926c-d2d6f9dfd584"
                  }
                ]
        default:
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Failed to get rulesets.",
                "status": 500,
                "detail": "Database operation failed."
              }
    post:
      tags:
        - rulesets
      summary: Create a new Ruleset.
      description: Create a new Ruleset comprised of the Ruleset name and a CSV file containing the rules.
      operationId: createRuleset
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateRuleset'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RulesetMetadata'
              example:
                {
                  "id": "e0399094-0b34-4ccd-bcf7-a904293594c5",
                  "rulesetName": "ruleset1",
                  "uri": "v1/rulesets/e0399094-0b34-4ccd-bcf7-a904293594c5"
                }
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/RuleErrorDetails'
              example: {
                "title": "Problems found in ruleset.",
                "status": 400,
                "detail": "Ruleset cannot contain any invalid MO types, attributes or values.",
                "ruleValidationErrors": [ {
                  "lineNumber": 5,
                  "errorType": "Invalid MO.",
                  "errorDetails": "MO not found in Managed Object Model.",
                  "additionalInfo": ""
                } ]
              }
        "413":
          $ref: '#/components/responses/413'
        "503":
          $ref: '#/components/responses/503'
        default:
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Failed to create ruleset.",
                "status": 500,
                "detail": "Database operation failed."
              }
...
  /v1/scopes:
    get:
      tags:
        - scopes
      summary: Get Metadata of all Scopes.
      description: Get metadata for all Scopes. An empty list is returned if no Scopes exist.
      operationId: listScopes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScopeMetadata'
              example:
                [
                  {
                    "id": "dedbfc30-41d5-46ff-ad39-ba36aebe2dbd",
                    "scopeName": "market1",
                    "uri": "v1/scopes/dedbfc30-41d5-46ff-ad39-ba36aebe2dbd"
                  },
                  {
                    "id": "6e57b373-cdb5-4d05-961f-17bef897ba9a",
                    "scopeName": "market2",
                    "uri": "v1/scopes/6e57b373-cdb5-4d05-961f-17bef897ba9a"
                  }
                ]
        default:
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Failed to get scopes.",
                "status": 500,
                "detail": "Database operation failed."
              }
    post:
      tags:
        - scopes
      summary: Create a new Scope.
      description: Create a new Scope comprised of the Scope name and a CSV file containing the ManagedElement FDNs.
      operationId: createScope
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateScope'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScopeMetadata'
              example:
                {
                  "id": "dedbfc30-41d5-46ff-ad39-ba36aebe2dbd",
                  "scopeName": "market1",
                  "uri": "v1/scopes/dedbfc30-41d5-46ff-ad39-ba36aebe2dbd"
                }
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "File does not match the csv format.",
                "status": 400,
                "detail": "Fill in all required fields with the appropriate information."
              }
        "413":
          $ref: '#/components/responses/413'
        default:
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Failed to create scope.",
                "status": 500,
                "detail": "Database operation failed."
              }
...
  /v1/jobs:
    get:
      tags:
        - jobs
      summary: Return all the Jobs
      description: Get all the Jobs defined in EACC
      operationId: getJobs
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'
              example:
                [
                  {
                    "jobName": "job_market1",
                    "scopeName": "market1",
                    "schedule": "0 38 14 19 2 *",
                    "rulesetName": "ruleset1"
                  },
                  {
                    "jobName": "job_market2",
                    "scopeName": "market2",
                    "schedule": "0 50 6 20 2 *",
                    "rulesetName": "ruleset2"
                  }
                ]
        default:
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Failed to get job(s).",
                "status": 500,
                "detail": "Database operation failed."
              }
    post:
      tags:
        - jobs
      summary: Create a Job
      description: Create a Job in EACC defining the schedule, rules and Scope of the configuration consistency check.
      operationId: postJobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Job'
            example:
              {
                "jobName": "job_market1",
                "scopeName": "market1",
                "schedule": "0 38 14 19 2 *",
                "rulesetName": "ruleset1"
              }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
              example:
                {
                  "jobName": "job_market1",
                  "scopeName": "market1",
                  "schedule": "0 38 14 19 2 *",
                  "rulesetName": "ruleset1"
                }
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Validation failed.",
                "status": 400,
                "detail": "Invalid jobName: job already exists"
              }
        default:
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Failed to create job.",
                "status": 500,
                "detail": "Database operation failed."
              }
...
components:
...
  schemas:
    RulesetMetadata:
      type: object
      required: [ id, rulesetName, uri ]
      properties:
        id:
          type: string
          description: The Ruleset ID.
        rulesetName:
          type: string
          description: The Ruleset name.
        uri:
          type: string
          description: The URI of the Ruleset.
    CreateRuleset:
      type: object
      required: [ rulesetName, fileName ]
      properties:
        rulesetName:
          type: string
          description: Use lower case alphanumeric, underscores and dashes up to a maximum of 100 characters.
        fileName:
          type: string
          format: binary
          description: |
            A CSV file containing the rules. See the User Guide for information on the format of this file.
...
    ErrorDetails:
      description: Provides additional information in an error response.
      type: object
      required: [ title, status, detail ]
      properties:
        title:
          $ref: '#/components/schemas/ErrorTitle'
        status:
          $ref: '#/components/schemas/ErrorStatus'
        detail:
          $ref: '#/components/schemas/ErrorDetail'
    RuleErrorDetails:
      description: Provides additional information in an error response.
      type: object
      required: [ title, status, detail ]
      properties:
        title:
          $ref: '#/components/schemas/ErrorTitle'
        status:
          $ref: '#/components/schemas/ErrorStatus'
        detail:
          $ref: '#/components/schemas/ErrorDetail'
        ruleValidationErrors:
          type: array
          description: |
            List of errors found in the Ruleset file being uploaded. Contains an empty list if the 
            issue is not a problem with the content of the file.
          items:
            type: object
            required: [ lineNumber, errorType, errorDetails, additionalInfo ]
            properties:
              lineNumber:
                type: integer
                format: int64
                description: The line number in the CSV file in which the error is located.
              errorType:
                type: string
                description: Slogan identifying the type of error.
              errorDetails:
                type: string
                description: Human-readable explanation of the error.
              additionalInfo:
                type: string
                description: Additional information if relevant otherwise an empty string.
...